Move ToopherIframe tests into own file and cleanup
