+++ test/test_toopher_api.php
@@ -26,45 +26,10 @@ use Rhumsaa\Uuid\Uuid;
 
 class ToopherApiTests extends PHPUnit_Framework_TestCase {
 
-    const OAUTH_NONCE = '12345678';
-    const IFRAME_KEY = 'abcdefg';
-    const IFRAME_SECRET = 'hijklmnop';
-    const REQUEST_TOKEN = 's9s7vsb';
-
-    public static function getOauthTimestamp()
-    {
-      return mktime(0, 16, 40, 1, 1, 1970);
-    }
-
-    public static function getOauthNonce()
-    {
-      return self::OAUTH_NONCE;
-    }
-
-    public static function getIframeKey()
-    {
-      return self::IFRAME_KEY;
-    }
-
-    public static function getIframeSecret()
-    {
-      return self::IFRAME_SECRET;
-    }
-
-    public static function getRequestToken()
-    {
-      return self::REQUEST_TOKEN;
-    }
-
     protected function setUp()
     {
       date_default_timezone_set('UTC');
       $this->mock = new HTTP_Request2_Adapter_Mock();
-    }
-
-    public function getToopherIframe()
-    {
-      return new ToopherIframe($this->getIframeKey(), $this->getIframeSecret(), 'https://api.toopher.test/v1/');
     }
 
     public function compareToDefaultUserTerminal($userTerminal)
@@ -559,164 +524,6 @@ class ToopherApiTests extends PHPUnit_Framework_TestCase {
       $action = new Action(["id" => "1", "name" => "action"]);
       $this->assertTrue($action->id == '1', 'bad action id');
       $this->assertTrue($action->name == 'action', 'bad action name');
-    }
-
-    public function testToopherIframeGetAuthenticationUrl()
-    {
-        $toopherIframe = $this->getToopherIframe();
-        $toopherIframe->setTimestampOverride($this->getOauthTimestamp());
-        $toopherIframe->setNonceOverride($this->getOauthNonce());
-        $expectedUrl = "https://api.toopher.test/v1/web/authenticate?v=2&username=jdoe&reset_email=jdoe%40example.com&action_name=Log+In&session_token=s9s7vsb&requester_metadata=None&expires=1300&oauth_consumer_key=abcdefg&oauth_signature_method=HMAC-SHA1&oauth_nonce=12345678&oauth_timestamp=1000&oauth_version=1.0&oauth_signature=YN%2BkKNTaoypsB37fsjvMS8vsG5A%3D";
-        $authenticationUrl = $toopherIframe->getAuthenticationUrl('jdoe', 'jdoe@example.com', $this->getRequestToken());
-        $this->assertTrue($authenticationUrl == $expectedUrl, 'authentication url is incorrect');
-    }
-
-    public function testToopherIframeGetAuthenticationUrlWithExtras()
-    {
-      $extras = array("allow_inline_pairing" => "false");
-      $toopherIframe = $this->getToopherIframe();
-      $toopherIframe->setTimestampOverride($this->getOauthTimestamp());
-      $toopherIframe->setNonceOverride($this->getOauthNonce());
-      $expectedUrl = "https://api.toopher.test/v1/web/authenticate?v=2&username=jdoe&reset_email=jdoe%40example.com&action_name=it+is+a+test&session_token=s9s7vsb&requester_metadata=None&expires=1300&allow_inline_pairing=false&oauth_consumer_key=abcdefg&oauth_signature_method=HMAC-SHA1&oauth_nonce=12345678&oauth_timestamp=1000&oauth_version=1.0&oauth_signature=W%2F2dcdsVc7YgdSCZuEo8ViHLlOo%3D";
-      $authenticationUrl = $toopherIframe->getAuthenticationUrl('jdoe', 'jdoe@example.com', $this->getRequestToken(), 'it is a test', 'None', $extras);
-      $this->assertTrue($authenticationUrl == $expectedUrl, 'authentication url is incorrect');
-    }
-
-    public function testToopherIframeGetUserManagementUrl()
-    {
-      $toopherIframe = $this->getToopherIframe();
-      $toopherIframe->setTimestampOverride($this->getOauthTimestamp());
-      $toopherIframe->setNonceOverride($this->getOauthNonce());
-      $expectedUrl = "https://api.toopher.test/v1/web/manage_user?v=2&username=jdoe&reset_email=jdoe%40example.com&expires=1300&oauth_consumer_key=abcdefg&oauth_signature_method=HMAC-SHA1&oauth_nonce=12345678&oauth_timestamp=1000&oauth_version=1.0&oauth_signature=NjwH5yWPE2CCJL8v%2FMNknL%2BeTpE%3D";
-      $userManagementUrl = $toopherIframe->getUserManagementUrl('jdoe', 'jdoe@example.com');
-      $this->assertTrue($userManagementUrl == $expectedUrl, 'user management url is incorrect');
-    }
-
-    public function testToopherIframeGetUserManagementUrlWithExtras()
-    {
-      $extras = array("ttl" => "100");
-      $toopherIframe = $this->getToopherIframe();
-      $toopherIframe->setTimestampOverride($this->getOauthTimestamp());
-      $toopherIframe->setNonceOverride($this->getOauthNonce());
-      $expectedUrl = "https://api.toopher.test/v1/web/manage_user?v=2&username=jdoe&reset_email=jdoe%40example.com&expires=1100&oauth_consumer_key=abcdefg&oauth_signature_method=HMAC-SHA1&oauth_nonce=12345678&oauth_timestamp=1000&oauth_version=1.0&oauth_signature=sV8qoKnxJ3fxfP6AHNa0eNFxzJs%3D";
-      $userManagementUrl = $toopherIframe->getUserManagementUrl('jdoe', 'jdoe@example.com', $extras);
-      $this->assertTrue($userManagementUrl == $expectedUrl, 'user management url is incorrect');
-    }
-
-    public function testToopherIframeValidatePostbackWithGoodSignatureIsSuccessful()
-    {
-      $toopherIframe = new ToopherIframe('abcdefg', 'hijklmnop', 'https://api.toopher.test/v1/');
-      $toopherIframe->setTimestampOverride($this->getOauthTimestamp());
-      $data = array(
-        'foo' => array('bar'),
-        'timestamp' => array($this->getOauthTimestamp()),
-        'session_token' => array('s9s7vsb'),
-        'toopher_sig' => array('6d2c7GlQssGmeYYGpcf+V/kirOI=')
-      );
-      try {
-        $toopherIframe->validatePostback($data, 's9s7vsb', 5);
-      } catch (Exception $e) {
-        $this->fail('Valid signature, timestamp, and session token did not return validated data');
-      }
-    }
-
-    /**
-     * @expectedException         SignatureValidationError
-     * @expectedExceptionMessage  Computed signature does not match
-     */
-    public function testToopherIframeValidatePostbackWithBadSignatureFails()
-    {
-      $toopherIframe = new ToopherIframe('abcdefg', 'hijklmnop', 'https://api.toopher.test/v1/');
-      $toopherIframe->setTimestampOverride($this->getOauthTimestamp());
-      $data = array(
-        'foo' => array('bar'),
-        'timestamp' => array(mktime(0, 16, 40, 1, 1, 1970)),
-        'session_token' => array('s9s7vsb'),
-        'toopher_sig' => array('invalid')
-      );
-      $toopherIframe->validatePostback($data, 's9s7vsb', 5);
-    }
-
-    /**
-     * @expectedException         SignatureValidationError
-     * @expectedExceptionMessage  TTL Expired
-     */
-    public function testToopherIframeValidatePostbackWithExpiredSignatureFails()
-    {
-      $toopherIframe = new ToopherIframe('abcdefg', 'hijklmnop', 'https://api.toopher.test/v1/');
-      $toopherIframe->setTimeStampOverride(mktime(0, 16, 40, 2, 1, 1970));
-      $data = array(
-        'foo' => array('bar'),
-        'timestamp' => array($this->getOauthTimestamp()),
-        'session_token' => array('s9s7vsb'),
-        'toopher_sig' => array('6d2c7GlQssGmeYYGpcf+V/kirOI=')
-      );
-      $toopherIframe->validatePostback($data, 's9s7vsb', 5);
-    }
-
-    /**
-    * @expectedException        SignatureValidationError
-    * @expectedExceptionMessage Session token does not match expected value
-    */
-    public function testToopherIframeValidatePostbackWithInvalidSessionTokenFails()
-    {
-      $toopherIframe = new ToopherIframe('abcdefg', 'hijklmnop', 'https://api.toopher.test/v1/');
-      $toopherIframe->setTimeStampOverride($this->getOauthTimestamp());
-      $data = array(
-        'foo' => array('bar'),
-        'timestamp' => array(mktime(0, 16, 40, 1, 1, 1970)),
-        'session_token' => array('invalid token'),
-        'toopher_sig' => array('6d2c7GlQssGmeYYGpcf+V/kirOI=')
-      );
-      $toopherIframe->validatePostback($data, 's9s7vsb', 5);
-    }
-
-    /**
-    * @expectedException        SignatureValidationError
-    * @expectedExceptionMessage Missing required keys: timestamp
-    */
-    public function testToopherIframeValidatePostbackMissingTimestampFails()
-    {
-      $toopherIframe = new ToopherIframe('abcdefg', 'hijklmnop', 'https://api.toopher.test/v1/');
-      $toopherIframe->setTimeStampOverride($this->getOauthTimestamp());
-      $data = array(
-        'foo' => array('bar'),
-        'session_token' => array('s9s7vsb'),
-        'toopher_sig' => array('6d2c7GlQssGmeYYGpcf+V/kirOI=')
-      );
-      $toopherIframe->validatePostback($data, 's9s7vsb', 5);
-    }
-
-    /**
-    * @expectedException        SignatureValidationError
-    * @expectedExceptionMessage Missing required keys: toopher_sig
-    */
-    public function testToopherIframeValidatePostbackMissingSignatureFails()
-    {
-      $toopherIframe = new ToopherIframe('abcdefg', 'hijklmnop', 'https://api.toopher.test/v1/');
-      $toopherIframe->setTimeStampOverride($this->getOauthTimestamp());
-      $data = array(
-        'foo' => array('bar'),
-        'session_token' => array('s9s7vsb'),
-        'timestamp' => mktime(0, 16, 40, 1, 1, 1970)
-      );
-      $toopherIframe->validatePostback($data, 's9s7vsb', 5);
-    }
-
-    /**
-    * @expectedException        SignatureValidationError
-    * @expectedExceptionMessage Missing required keys: session_token
-    */
-    public function testToopherIframeValidatePostbackMissingSessionTokenFails()
-    {
-      $toopherIframe = new ToopherIframe('abcdefg', 'hijklmnop', 'https://api.toopher.test/v1/');
-      $toopherIframe->setTimeStampOverride($this->getOauthTimestamp());
-      $data = array(
-        'foo' => array('bar'),
-        'timestamp' => array($this->getOauthTimestamp()),
-        'toopher_sig' => array('6d2c7GlQssGmeYYGpcf+V/kirOI=')
-      );
-      $toopherIframe->validatePostback($data, 's9s7vsb', 5);
     }
 
     /**
